# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    email: p.korotkevitch@gmail.com
    name: Pavel Karatkevich
    url: https://www.linkedin.com/in/pavel-karatkevich-236461178/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soundproof service
  version: "2.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login form which received login/password, generates JWT token and
        returns a login response (user info)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Login form
  /auth/register/:
    post:
      consumes:
      - application/json
      description: Register a user by passing a User Registration Request via the
        context
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Register a user
  /user/profile:
    get:
      consumes:
      - application/json
      description: Received login request via context, checks JWT token and retrieves
        user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: type
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Gets a user
    put:
      consumes:
      - application/json
      description: Update user info by passing a User Update Request via the context.
        Parses signature and signed string from Metamask and stores a Metamask public
        address into the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update user info
swagger: "2.0"